clean neoscan of all reference to monitor
